chore(tracex): rename transaction-tracing to tracex
