+++ Cargo.toml
@@ -55,7 +55,7 @@ base-reth-cli = { path = "crates/cli" }
 base-reth-test-utils = { path = "crates/test-utils" }
 base-reth-flashblocks = { path = "crates/flashblocks" }
 base-reth-rpc = { path = "crates/rpc" }
-base-reth-transaction-tracing = { path = "crates/transaction-tracing" }
+base-tracex = { path = "crates/tracex" }
 base-reth-runner = { path = "crates/runner" }
 
 # base/tips
