+++ crates/tracex/README.md
@@ -1,0 +1,4 @@
+## Tracex
+
+Base's execution extension for tracking transaction lifecycle events
+from mempool insertion to block inclusion.
