+++ crates/tracex/src/lib.rs
@@ -3,7 +3,12 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-mod tracing;
-pub use tracing::transaction_tracing_exex;
+/// Domain events tracked by tracex.
+pub mod events;
+mod exex;
+mod tracker;
 
-mod types;
+pub use crate::{
+    events::{EventLog, Pool, TxEvent},
+    exex::tracex_exex,
+};
