--- crates/transaction-tracing/README.md
@@ -1,6 +1,0 @@
-## Transaction Tracing
-
-Utilities for transaction tracing.
-
-This crate provides an ExEx for tracking transaction lifecycle events
-from mempool insertion to block inclusion.
